requirements.txt
pandas
gspread
google-api-python-client
PyDrive
requests
requests-oauthlib
oauth2client
discord.py
urllib
